% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table-Functions.R
\name{bi_table}
\alias{bi_table}
\title{Table of Biased Items}
\usage{
bi_table(dif.analysis)
}
\arguments{
\item{dif.analysis}{Output from \code{\link[WBdif]{dif_analysis}}}
}
\value{
a \code{data.frame} with biased items in rows and DIF methods in columns
}
\description{
Compiles table of biased items identified by selected DIF methods
}
