% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-Functions.R
\name{bias_plot}
\alias{bias_plot}
\title{Plotting Treatment Effects}
\usage{
bias_plot(dif.analysis)
}
\arguments{
\item{dif.analysis}{an object returned from \code{dif_analysis}}
}
\description{
Functions to visualize treatment effect and bias magnitude
}
