% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newPlot-Functions.R
\name{bias_plots}
\alias{bias_plots}
\title{Plots of test bias}
\usage{
bias_plots(dif.models)
}
\arguments{
\item{dif.models}{Output from \code{\link[WBdif]{dif_models}}.}
}
\value{
A \code{list} of two \code{ggplot} objects.
}
\description{
Plots test characteristic curves from a two-group IRT model and their difference.
}
\details{
This is a \code{ggplot2} wrapper for plotting test characteristic curves (TCCs). It calls \code{\link[mirt]{expected.test}} and plots the resulting TCCs for both levels of \code{dif.group.id}, as well as their difference.
}
