% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Analysis.R
\name{dif_analysis}
\alias{dif_analysis}
\title{Differential item functioning}
\usage{
dif_analysis(
  measure.data,
  dif.group,
  methods = c("loess", "MH", "logistic", "IRT"),
  score.type = c("Rest", "Total"),
  match.bins = NULL
)
}
\arguments{
\item{measure.data}{data frame of dichotomous item responses with subjects in rows
and items in columns}

\item{dif.group}{factor or character vector of group membership for which DIF
is evaluated. If a character vector, this will be transformed to a [base::factor()].}

\item{methods}{Character vector with one or more of the four available methods:
"loess", "MH", "logistic", and "IRT".
The default is all four methods.}

\item{score.type}{character indicating whether the total summed score ("Total") or the
summed score excluding the item under investigation ("Rest") should be used for the
matching variable in the loess, MH, or logistic regression methods.}
}
\value{
a list with DIF results from each selected method
}
\description{
Evaluates differential item functioning (DIF) using loess, Mantel-Haenszel (MH),
logistic regression, and item response theory (IRT) approaches.
}
\details{
This function primarily serves as a wrapper around the method specific functions, but
also identifies and reports items with no variance (e.g., all 0 responses) or no
variance within the levels of `dif.group`. The "loess" method produces loess curves
for every item in `measure.data`. Items with no variance are removed from
`measure.data` for the "MH", "logistic", and "IRT" methods because response
variance is a prerequisite for DIF. Additionally, items with no variance within
a `dif.group` level are removed from the "IRT" method because these lead to
the models being under-identified.
}
