% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newData-Prep.R
\name{dif_data_prep}
\alias{dif_data_prep}
\title{Data prep for \code{WBdif} functions.}
\usage{
dif_data_prep(
  item.data,
  tx.group.id = NULL,
  dif.group.id = tx.group.id,
  cluster.id = NULL,
  std.group = NULL,
  na.to.0 = FALSE
)
}
\arguments{
\item{item.data}{A \code{data.frame} of item responses with subjects in rows
and items in columns. Data can be binary or ordered categorical.}

\item{tx.group.id}{A \code{vector} of length \code{nrow(item.data)} indicating the treatment groups. Must have exactly two unique value.}

\item{dif.group.id}{A \code{vector} of length \code{nrow(item.data)} indicating the DIF groups. Must have exactly two unique value. See details for discussion of default value.}

\item{cluster.id}{An optional \code{vector} of length \code{nrow(item.data)} indicating the primary sampling unit in a multi-stage / clustered sampling design -- used to adjust effect sizes and their standard errors.}

\item{std.group}{An optional value of \code{tx.group.id} that identifies the group whose standard deviation will be used to standardize the treatment effect.}

\item{na.to.0}{After removing empty rows, should remaining NAs in \code{item.data} be converted to 0? Default is FALSE.}
}
\value{
A named \code{list} containing the pre-processed inputs and item flags.
}
\description{
Data pre-processing and formatting; needs to be called before using other \code{WBdif} functions.
}
\details{
This function saves the input data in a format used by other \code{WBdif} functions. The relation between \code{tx.group.id} and \code{dif.group.id} is especially important. If only one is supplied (or these are equal), then \code{\link[WBdif]{effect_robustness}} evaluates the unconditional treatment effects; if both are specified and they are not equal, then \code{\link[WBdif]{effect_robustness}} evaluates treatment effects conditional on the DIF groups (e.g., conditional on gender).

This function also runs a number of pre-processing steps:
\itemize{
  \item Drop cases with empty rows \code{item.data} (i.e., rows with all NAs) or with \code{NA} for \code{dif.group.id}, \code{tx.group.id}, or \code{cluster.id}.
  \item Flag items with no variance.
  \item Flag items with different number of response categories in the different DIF groups.
  \item Flag items with more than two response categories.
}
}
\examples{
data("mdatlang")

dif.data <- dif_data_prep(item.data = mdatlang`[`5:ncol(mdatlang)`]`,
                             tx.group.id = mdatlang$treated,
                             dif.group.id = mdatlang$gender,
                             cluster.id = mdatlang$clusterid,
                             na.to.0 = TRUE)

}
