% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic-Functions.R
\name{dif_logistic}
\alias{dif_logistic}
\title{Logistic regression DIF method}
\usage{
dif_logistic(item.data, dif.group.id, match.type, match.scores)
}
\arguments{
\item{item.data}{A \code{data.frame} of item responses with subjects in rows
and items in columns.}

\item{dif.group.id}{A factor vector of group membership for which DIF is evaluated.}

\item{match.type}{A character indicating whether \code{match.scores} is a
total summed score ("Total") or the summed score excluding the item under
investigation ("Rest").}

\item{match.scores}{A numeric vector (if \code{match.type = "Total"}) or
\code{list} of \code{ncol(item.data)} numeric vectors (if \code{match.type = "Rest"})
of match scores used as a predictor in the logistic regression.}
}
\value{
list containing
\itemize{
  \item DIF model comparisons
  \item item-level DIF tests
  \item integer vector of the items showing DIF (i.e., biased items)
  \item type of DIF
  }
}
\description{
Conducts DIF analysis with logistic regression
}
\details{
First conducts an omnibus test of DIF by comparing fit of no DIF, uniform DIF, and non-uniform DIF logistic
regression models. The no DIF model regresses the dichotomous item responses from \code{item.data}
on item, \code{match.scores}, and the two-way interaction. The uniform DIF model adds \code{dif.group.id}
and the interaction with item while the non-uniform model adds the three-way interaction. If DIF is
detected through the model comparisons, the specific item(s) with DIF are
identified in a two-stage process - initial detection stage and refinement stage.
}
