% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH-Functions.R
\name{dif_mh}
\alias{dif_mh}
\title{Mantel-Haenszel DIF method}
\usage{
dif_mh(
  item.data,
  dif.groups,
  match.type,
  match.scores,
  match.bins = NULL,
  exclude = NULL,
  stop = TRUE
)
}
\arguments{
\item{item.data}{A \code{data.frame} of dichotomous item responses with subjects in rows and items in columns.}

\item{dif.groups}{factor vector of group membership for which DIF is evaluated.}

\item{match.type}{For the loess, MH, and logistic methods, a character indicating whether a
total summed score (\code{"Total"}; default) or the summed score excluding the item under
investigation (\code{"Rest"}) should be used as the stratifying variable.}

\item{match.scores}{A numeric vector (if \code{match.type = "Total"}) or \code{list} of \code{ncol(item.data)}
numeric vectors (if \code{match.type = "Rest"}) of match scores used as the stratifying variable in the MH procedure.}

\item{match.bins}{For MH, an optional vector of bin sizes for stratifying the \code{match.type} score.
This is passed to the \code{probs} argument of \code{stats::quantile}. See examples for usage.}

\item{exclude}{character vector of column names in \code{item.data} to exclude from DIF analysis, such as the anchor items or (currently) polytomous items.}

\item{stop}{If \code{FALSE}, when possible, errors will return \code{NA} for the item under investigation rather than stop the analysis completely (default is \code{TRUE}.}
}
\value{
A list containing
\itemize{
\item a \code{data.frame} with MH results for each item in \code{item.data}
\item an integer vector of the items showing DIF (i.e., biased items)
}
}
\description{
Conducts DIF analysis with the Mantel-Haenszel method
}
\details{
\code{run_mh} conducts a Mantel-Haenszel test via \code{\link[stats]{mantelhaen.test}} for each
item in \code{item.data}, except those specified in \code{exclude}, grouped by \code{dif.groups} and stratified by the scores supplied
to \code{match.scores}, which, if specified, are binned by \code{match.bins}. Biased items
are identified through a two-step process of initial and refinement phases.
}
