% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Report.R
\name{dif_report}
\alias{dif_report}
\title{Generate report of DIF analysis}
\usage{
dif_report(
  dif.analysis,
  dataset.name,
  measure.name,
  dif.group.name,
  bias.method = "IRT",
  irt.scoring = "WLE",
  tx.group = NULL
)
}
\arguments{
\item{dif.analysis}{an object returned from `dif_analysis`}

\item{dataset.name}{character; name of the dataset where the measure and
item responses came from. This is printed in the report title and summary.
For World Bank data, this will typically be a country.}

\item{measure.name}{character; name of the measure being evaluated for DIF.
This is printed in the report title and summary.}

\item{dif.group.name}{character; name of the group used in the DIF analysis.}

\item{bias.method}{options are "MH", "logistic", or "IRT" (default).
Which method in the `dif.analysis` object should the biased items, if detected, be
extracted?}

\item{irt.scoring}{factor score estimation method, which is passed to
[mirt::fscores()]. Default is "WLE". See [mirt::fscores()] documentation for
other options.}

\item{tx.group}{When `NULL` (default) unconditional treatment effects are assessed
for robustness with the assumption that the treatment indicator was used as the
grouping variable in `dif_analysis`. If treatment indicator was not the grouping
variable in `dif_analysis`, the factor vector for treatment indicator needs to be
supplied here. Robustness of the conditional treatment effects is then assessed.}
}
\value{
Uses the template "Bias_Correction_Report.Rmd" to produce a
report summarizing whether any items on `measure.name` are biased with
respect to `dif.group.name`, and, if so, to what extent this affects
treatment contrasts on `measure.name`.
}
\description{
Produces a report summarizing an analysis of measurement bias for a given
scale and grouping variable
}
