% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot-Functions.R
\name{effects_plot}
\alias{effects_plot}
\title{Forest plots of effect sizes}
\usage{
effects_plot(effects)
}
\arguments{
\item{effects}{Output from \code{\link[WBdif]{effect_robustness}}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plots standardized treatment effects and their confidence intervals.
}
\details{
This is a \code{ggplot2} wrapper for producing forest plots of effect sizes.
}
