% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table-Functions.R
\name{effects_table}
\alias{effects_table}
\title{Table of Treatment Effect Estimates}
\usage{
effects_table(effects, alphas = NULL)
}
\arguments{
\item{effects}{Output from \code{\link[WBdif]{effect_robustness}}}

\item{alphas}{(optional) \code{vector} of score reliabilities (e.g., alpha, omega) to append to the table.}
}
\value{
a \code{data.frame} of treatment effect estimates and their standard errors.
}
\description{
Compiles table of treatment effect estimates
}
