% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est-SMD.R
\name{est_smd}
\alias{est_smd}
\title{Standardized Mean Difference}
\usage{
est_smd(
  outcome = NULL,
  groups = NULL,
  m1 = NULL,
  m2 = NULL,
  sdp = NULL,
  sd1 = NULL,
  sd2 = NULL,
  n1 = NULL,
  n2 = NULL,
  hedges.g = FALSE,
  cluster.n = NULL,
  icc = NULL
)
}
\arguments{
\item{outcome}{numeric vector of outcome values}

\item{groups}{factor or character vector with 2 levels indicating group membership}

\item{m1}{mean for group 1}

\item{m2}{mean for group 2}

\item{sdp}{user-specified divisor for estimating the smd
(e.g., population sd, pooled sd)}

\item{sd1}{standard deviation for group 1}

\item{sd2}{standard deviation for group 2}

\item{n1}{sample size for group 1}

\item{n2}{sample size for group 2}

\item{hedges.g}{unbiased estimation of standardized mean difference?}

\item{cluster.n}{average cluster size}

\item{icc}{intraclass correlation - proportion of total variance that is between
cluster variance}
}
\value{
If both \code{n1} and \code{n2} are supplied the standardized mean difference
and its variance are returned in a data.frame. Otherwise only the estimated
standardized mean difference is returned as a numeric value.
}
\description{
Estimates the standardized mean difference with or without adjusting for clustering
}
\details{
The standardized mean difference can be calculated from the raw data
(\code{outcome} and \code{groups}) or from the sample statistics.
If raw data is supplied the pooled standard deviation is used as the divisor
unless \code{sdp} is specified.

Both \code{cluster.n} and \code{icc} must be supplied to adjust for clustering.
Otherwise, the unadjusted estimate and standard error are returned without a warning.
The cluster-adjustment assumes the average cluster size is equal in group 1
and group 2.
}
\references{
Cohen, J. (1988). \emph{Statistical power analysis for the behavioral sciences}
(2nd ed.). Hillsdale, NJ: Lawrence Erlbaum.
Hedges, L. V. (1981). Distribution theory for Glass’s estimator of effect size
and related estimators. \emph{Journal of Educational Statistics, 6}(2), 107–128.
Hedges, L. V. (2007). Effect sizes in cluster-randomized designs.
\emph{Journal of Educational and Behavioral Statistics, 32}(4), 341-370.
}
