% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table-Functions.R
\name{extract_biased_items}
\alias{extract_biased_items}
\title{Extract Biased Items}
\usage{
extract_biased_items(dif.analysis)
}
\arguments{
\item{dif.analysis}{Output from \code{\link[DIFreport]{dif_analysis}}}
}
\value{
a \code{list} of vectors of column locations in \code{item.data} for the biased items identified by each DIF method.
}
\description{
Extracts list of biased items from a \code{dif_analysis} object
}
