% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hedges2007.R
\name{get_avg_cluster_n}
\alias{get_avg_cluster_n}
\title{Compute average cluster size}
\usage{
get_avg_cluster_n(tx.group.id, cluster.id)
}
\arguments{
\item{tx.group.id}{A \code{vector} of indicating the treatment groups.}

\item{cluster.id}{An optional \code{vector} of \code{length(group.id)} indicating the primary sampling unit in a multi-stage / clustered sampling design.}
}
\value{
The average cluster size; returns 0 if \code{is.null(cluster.id)}.
}
\description{
Helper function for \code{hedges2007} that computes Equation 19 of Hedges (2007).
}
