% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hedges2007.R
\name{get_icc}
\alias{get_icc}
\title{Estimate the intraclass correlation (ICC)}
\usage{
get_icc(outcome, cluster.id)
}
\arguments{
\item{outcome}{A numeric \code{vector} containing the outcome variable.}

\item{cluster.id}{An optional \code{vector} of \code{length(outcome)} indicating the primary sampling unit in a multi-stage / clustered sampling design.}
}
\value{
The ICC; returns 0 if \code{is.null(cluster.id)}.
}
\description{
Helper function for \code{hedges2007} that requires \code{lme4}.
}
