% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic-Functions.R
\name{get_logistic}
\alias{get_logistic}
\alias{run_global_logistic}
\alias{run_item_logistic}
\title{Logistic regression DIF method}
\usage{
get_logistic(scale.data, dif.group, score.type, match)

run_global_logistic(scale.data, dif.group, match)

run_item_logistic(item.data, dif.type)
}
\arguments{
\item{scale.data}{data frame of dichotomous item responses with subjects in rows
and items in columns}

\item{dif.group}{factor vector of group membership for which DIF is evaluated.}

\item{score.type}{character indicating whether `match` is a
total summed score ("Total") or the summed score excluding the item under
investigation ("Rest").}

\item{match}{list of `ncol(scale.data)` elements. Each element is a numeric vector
of match scores used as a predictor in the logistic regression.
In `get_logistic`, if `score.type` = "Total", `match` will be a single numeric vector.}

\item{item.data}{dataframe created in `run_global_logistic`}

\item{dif.type}{character indicating "No DIF", "uniform" DIF or "non-uniform" DIF as
determined in `run_global_logistic`.}
}
\value{
a four-element list containing 1) DIF model comparisons,
2) item-level DIF tests, 3) integer vector of the items showing DIF
(i.e., biased items), and 4) type of DIF
}
\description{
Conducts DIF analysis with logistic regression
}
\details{
`run_global_logistic` compares no DIF, uniform DIF, and non-uniform DIF logistic
regression models by regressing dichotomous item responses from `scale.data` on item,
`match`, and `dif.group`, along with interactions. If DIF is detected through the
model comparisons, the specific item(s) with DIF are identified via
`run_item_logistic`. `get_logistic` is a wrapper around the other functions that
organizes the initial and refinement phases of the DIF analysis and compiles
the results.
}
