% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH-Functions.R
\name{get_mh}
\alias{get_mh}
\alias{run_mh}
\title{Mantel-Haenszel DIF method}
\usage{
get_mh(scale.data, dif.group, score.type, match, match.bins = NULL)

run_mh(scale.data, dif.group, item, match, match.bins = NULL)
}
\arguments{
\item{scale.data}{data frame of dichotomous item responses with subjects in rows
and items in columns}

\item{dif.group}{factor vector of group membership for which DIF is evaluated.}

\item{score.type}{character indicating whether `match` is a
total summed score ("Total") or the summed score excluding the item under
investigation ("Rest").}

\item{match}{numeric vector used as the stratifying variable in the MH procedure.
In `get_mh`, if `score.type` = "Rest", `match` will be a list of vectors -
one for each item in `scale.data`.}

\item{item}{integer; item in `scale.data` under investigation for DIF}
}
\value{
A two-element list containing 1) a data frame with MH results for each item
in `scale.data` and 2) an integer vector of the items showing DIF (i.e., biased items).
}
\description{
Conducts DIF analysis with Mantel-Haenszel method
}
\details{
`run_mh` conducts a Mantel-Haenszel test via [stats::mantelhaen.test] for each
item in `scale.data` grouped by `dif.group` and stratified by the scores supplied
to `match`, which, if specified, are binned by `match.bins`.
`get_mh` is a wrapper around `run_mh` that organizes the initial and refinement
phases of the DIF analysis and compiles the results.
}
