% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robustness-Functions.R
\name{get_robustness}
\alias{get_robustness}
\title{Robustness of treatment effects}
\usage{
get_robustness(
  scale.data,
  dif.group,
  biased.items = NULL,
  no.var.items = numeric(),
  no.dif.mod = NULL,
  irt.scoring = "WLE",
  tx.group = NULL
)
}
\arguments{
\item{scale.data}{data frame of dichotomous item responses with subjects in rows
and items in columns}

\item{dif.group}{factor vector of group membership for which DIF is evaluated}

\item{biased.items}{numeric vector; location of biased items in `scale.data`}

\item{no.var.items}{numeric vector; location of items in `scale.data` with 0 variance}

\item{no.dif.mod}{`MultipleGroupClass` object from `mirt` where parameters are
constrained to be equal between levels of `dif.group`}

\item{irt.scoring}{factor score estimation method, which is passed to the
`method` argument of [mirt::fscores()]}

\item{tx.group}{factor vector for treatment indicator}

\item{m1}{mean for dif.group 1}

\item{m2}{mean for dif.group 2}

\item{sd1}{sd for dif.group 1}

\item{sd2}{sd for dif.group 2}

\item{n1}{sample size for dif.group 1}

\item{n2}{sample size for dif.group 2}

\item{score}{vector of scale scores}
}
\value{

}
\description{
Evaluates the robustness of treatment effects when DIF is detected
}
\details{

}
