% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hedges2007.R
\name{get_sd}
\alias{get_sd}
\title{Computes the pooled standard deviation}
\usage{
get_sd(outcome, group.id, std.group = NULL)
}
\arguments{
\item{outcome}{A numeric \code{vector} containing the outcome variable.}

\item{group.id}{A \code{vector} of \code{length(outcome)} indicating the treatment groups.}

\item{std.group}{An optional value of \code{tx.group.id} that identifies the group whose standard deviation will be used to standardize the treatment effect. If \code{NULL} (default), the pooled standard devaition is used.}
}
\value{
If \code{is.null(std.group)}, returns the pooled standard deviation; else returns the standard deviation of \code{std.group}.
}
\description{
Helper function for \code{hedges2007}.
}
