% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MDAT-Documentation.R
\docType{data}
\name{mdat}
\alias{mdat}
\title{Malawi Developmental Assessment Tool: Language}
\format{
A data frame with 2122 rows and 30 variables:
\describe{
  \item{clusterid}{cluster identifier}
  \item{id}{subject identifier}
  \item{treated}{treatment condition; "Tx", "Control"}
  \item{gender}{gender; "Male", "Female"}
  \item{l20 - l48}{dichotomous (1/0) item responses}
}
}
\source{
\url{https://www.worldbank.org/en/programs/sief-trust-fund/brief/effects-of-quality-improvement-strategies-on-early-childhood-development-in-community-based-childcare-centers-in-malawi}
}
\usage{
mdat
}
\description{
A dataset containing the 26 item MDAT language assessment and indicators for treatment condition and gender
}
\keyword{datasets}
