% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRT-Functions.R
\name{run_item_irt}
\alias{run_item_irt}
\title{IRT item tests for DIF}
\usage{
run_item_irt(
  global.irt,
  which.model = "nonuniform.mod",
  items2test,
  scheme = "add",
  Wald = TRUE
)
}
\arguments{
\item{global.irt}{an object returned from \code{\link[DIFreport]{run_global_irt}}}

\item{which.model}{the model in \code{global.irt} to use for testing DIF. Options
are \code{"no.dif.mod"} (default), \code{"uniform.mod"} or \code{"nonuniform.mod"}.}

\item{items2test}{numeric; the item to test for DIF, which is passed to the
items2test argument of \code{\link[mirt]{DIF}}.}
}
\value{
an object returned from \code{\link[mirt]{DIF}}
}
\description{
Test each item for DIF using IRT models
}
