% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Logistic-Functions.R
\name{run_item_logistic}
\alias{run_item_logistic}
\title{Single Item Logistic Models for DIF}
\usage{
run_item_logistic(item.data, dif.type)
}
\arguments{
\item{item.data}{subset of long format data returned from \code{\link[WBdif]{run_global_logistic}} for a single item}

\item{dif.type}{extracted from object returned from \code{\link[WBdif]{run_global_logistic}}}
}
\value{
a \code{data.frame} containing results of a likelihood ratio test
}
\description{
Internal function for dif_logistic
}
