% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loess-Functions.R
\name{run_loess}
\alias{run_loess}
\title{LOESS regression for DIF analysis}
\usage{
run_loess(item.data, dif.group.id, item, match.scores, pred.scores, nitems)
}
\arguments{
\item{item}{integer; item in \code{item.data} under investigation for DIF}

\item{pred.scores}{range of scores to use in \code{\link[stats]{predict}}
for generating the LOESS curves. The values are defined in \code{link[WBdif]{dif_loess}}.}

\item{nitem}{number of items in the scale. Defined in \code{link[WBdif]{dif_loess}}.}
}
\value{
a \code{data.frame}
}
\description{
Internal function for dif_loess
}
