% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-Functions.R
\name{sum_score}
\alias{sum_score}
\alias{extract_bi}
\alias{format_flex}
\title{Internal convenience functions}
\usage{
sum_score(scale.data, drops = NULL)

extract_bi(dif.analysis)

format_flex(df, bold.bias = "no", digits = 3)
}
\arguments{
\item{scale.data}{data frame of dichotomous item responses used to calculate the score}

\item{drops}{integer vector of item locations in \code{scale.data} to exclude from
score calculation}

\item{dif.analysis}{an object returned from \code{dif_analysis}}

\item{df}{a data frame}

\item{bold.bias}{character indicating whether results tables should bold instances
of DIF or not ("no", the default). If so, must specify whether the table
displays model comparisons ("model") or item-level tests ("item").}

\item{digits}{integer; number decimal places to print in report tables}
}
\value{
\code{sum_score} returns a numeric vector of summed scores
\code{extract_bi} returns list of biased items from each \code{dif.analysis} method
\code{format_flex} returns a \code{\link[flextable]{flextable}} object
}
\description{
Collection of convenience functions to improve automaticity and code readability
}
\details{
\code{sum_score} sums dichotomous item responses into a scale score
\code{extract_bi} extracts the biased items identified from each \code{dif.analysis} method.
If no biased items were detected, a character string indicated so is extracted.
\code{format_flex} converts the DIF results from a data frame to a
\code{\link[flextable]{flextable}} and formats the table for printing in the report.
}
