% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summarize-DIF.R
\name{summarize_dif}
\alias{summarize_dif}
\title{Summarize DIF Analysis Results}
\usage{
summarize_dif(
  dif.data,
  report.type = "dif.effects",
  report.title = NULL,
  measure.name = NULL,
  file.name = measure.name,
  dataset.name = NULL,
  methods = c("loess", "MH", "logistic", "IRT"),
  bias.method = "IRT",
  match.type = "Total",
  match.bins = NULL,
  irt.scoring = "WLE"
)
}
\arguments{
\item{dif.data}{The output of \code{\link[WBdif]{dif_data_prep}}.}

\item{report.type}{Produce report including both the DIF analysis results and treatment effect robustness checks ("dif.effects"; default), only DIF analysis ("dif.only"), or only treatment effect robustnes checks ("effects.only")?}

\item{report.title}{Title of report}

\item{measure.name}{Name of the measure being evaluated for DIF, which is used in the report.}

\item{file.name}{File name to create on disk. Defaults to \code{measure.name}. The file path can also be specified here. The file is saved to the working directory by default.}

\item{dataset.name}{(optional) name of the dataset where the measure and
item responses came from, which is printed in the report. For World Bank data, this will typically be a country.}

\item{bias.method}{From which method in \code{methods} should the biased items, if detected, be extracted? The options are "MH", "logistic", or "IRT" (default).}

\item{match.type}{Οne of \code{c("Total", "Rest")}. Determines whether the total score or rest score should be used as the stratifying variable for loess, MH, and logistic regression methods.}

\item{match.bins}{(optional) vector of bin sizes for stratifying the matching variable in
the MH method. This is passed to the \code{probs} argument of \code{stas::quantile}.}

\item{irt.scoring}{Factor score estimation method, which is passed to
\code{\link[mirt]{fscores}}. Default is "WLE". See \code{\link[mirt]{fscores}} documentation for other options.}
}
\description{
Conduct a DIF analysis and produce report summarizing results
}
\details{
This function is a wrapper around \code{\link[WBdif]{dif_analysis}}, \code{\link[WBdif]{dif_models}}, \code{\link[WBdif]{effect_robustness}}, and \code{\link[WBdif]{dif_report}} in order to simplify report production when desired.
}
\examples{
data("mdatlang")

# prep data
dif.data <- dif_data_prep(item.data = mdatlang`[`5:ncol(mdatlang)],
                          tx.group.id = mdatlang$treated,
                          dif.group.id = mdatlang$gender,
                          cluster.id = mdatlang$clusterid,
                          na.to.0 = TRUE)

summarize_dif(dif.data = dif.data, report.type = "dif.effects",
              report.title = "Gender DIF on MDAT Language",
              measure.name = "MDAT Language", file.name = "DIF-Gender-MDAT-Language")
}
