% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data-Prep.R
\name{wb_data_prep}
\alias{wb_data_prep}
\title{Preparing World Bank Data}
\usage{
wb_data_prep(data, wb.items, tx.group.name, dif.group.name = tx.group.name)
}
\arguments{
\item{data}{\code{data.frame} containing item responses and group indicators.}

\item{wb.items}{regex character string identifying item response columns in \code{data}}

\item{tx.group.name}{character; variable name of treatment indicator in \code{data}.}

\item{dif.group.name}{optional character; variable name in \code{data} for which DIF
is evaluated. Supply only if different from \code{tx.group.name}
Default is \code{tx.group.name}}
}
\value{
A named list where \code{measure.data} is the data frame of item responses
and \code{dif.group} is the vector of group membership for which DIF is evaluated.
If \code{dif.group} is not the treatment indicator, then the vector of treatment indicators
is returned in the element \code{tx.group}.
}
\description{
Function specifically for cleaning World Bank Data to run a DIF analysis and report
}
\details{
Currently, \code{tx.group.name} (and \code{dif.group.name} if supplied) must name
a factor with 2 (and only 2) levels
}
