% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Models.R
\name{dif_models}
\alias{dif_models}
\title{IRT models with biased items}
\usage{
dif_models(dif.analysis = NULL, dif.data = NULL, biased.items = "IRT")
}
\arguments{
\item{dif.analysis}{Output from \code{\link[WBdif]{dif_analysis}}}

\item{dif.data}{Output from \code{\link[WBdif]{dif_data_prep}}. Only required if \code{dif.analysis} is not provided.}

\item{biased.items}{If \code{dif.analysis} is specified, a character indicating the method from which to extract the biased items. Options are "IRT" (default), "logistic", "MH".
If \code{dif.data} is specified, the \code{vector} of column indices indicating the biased items in \code{dif.data$item.data}}
}
\value{
A named \code{list} containing the data and IRT models to estimate treatment effects.
}
\description{
Primarily an interim function between dif_analysis and effect_robustness
}
\details{
Estimates a 2PL IRT model with the parameters on biased items feed over groups - the DIF model.
The No DIF model, which constrains all item parameters to be equal across the groups, is either extracted from \code{dif.analysis} or estimated as well.
The IRT models are estimated using \code{\link[mirt]{multipleGroup}}.
}
