% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Models.R
\name{dif_models}
\alias{dif_models}
\title{IRT models with biased items}
\usage{
dif_models(dif.analysis = NULL, dif.data = NULL, biased.items = "IRT")
}
\arguments{
\item{dif.analysis}{Output from \code{\link[WBdif]{dif_analysis}}}

\item{dif.data}{Output from \code{\link[WBdif]{dif_data_prep}}. Only required if \code{dif.analysis} is not provided.}

\item{biased.items}{If \code{dif.analysis} is specified, a character indicating the method from which to extract the biased items. Options are "IRT" (default), "logistic", "MH".
If \code{dif.data} is specified, the \code{vector} of column indices indicating the biased items in \code{dif.data$item.data}}
}
\value{
A named \code{list} containing the data and IRT models to estimate treatment effects.
}
\description{
Primarily an interim function between dif_analysis and effect_robustness
}
\details{
Estimates a 2PL IRT model with the parameters on biased items feed over groups - the DIF model.
The No DIF model, which constrains all item parameters to be equal across the groups, is either extracted from \code{dif.analysis} or estimated as well.
The IRT models are estimated using \code{\link[mirt]{multipleGroup}}.
}
\examples{
data("mdat")

# prep data
dif.data <- dif_data_prep(item.data = mdat`[`5:ncol(mdat)],
                             dif.group.id = mdat$gender,
                             na.to.0 = TRUE)

# With biased items identified via DIF analysis
dif.analysis <- dif_analysis(dif.data = dif.data,
                             methods =  c("MH", "IRT"),
                             match.type = "Rest",
                             match.bins = seq(0, 1, by = .1))
dif.models <- dif_models(dif.analysis = dif.analysis, biased.items = "MH")

# With user-specified biased items
dif.models <- dif_models(dif.data = dif.data, biased.items = c(1, 5, 7))

}
