% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Prep.R
\name{dif_prep}
\alias{dif_prep}
\title{Preparing data for DIF analysis}
\usage{
dif_prep(
  measure.data,
  dif.group,
  tx.group = dif.group,
  clusters = NULL,
  na0 = FALSE
)
}
\arguments{
\item{measure.data}{data frame of item responses with subjects in rows
and items in columns.}

\item{dif.group}{vector of group membership by which DIF is evaluated.}

\item{tx.group}{vector indicating treatment condition unless treatment indicator was
already supplied to \code{dif.group}.}

\item{clusters}{vector of cluster membership.}

\item{na0}{after removing empty rows, should remaining NAs be converted to 0?
Default is FALSE.}
}
\value{
list containing \code{measure.data} and \code{dif.group} (if supplied, also
\code{tx.group} and \code{clusters}) after consistent handling of missing data.
}
\description{
Handle missing data to run a DIF analysis and report
}
\details{
Identifies empty rows in measure.data (i.e., rows with all NAs) and rows with NA for
the \code{dif.group} and, if supplied, \code{tx.group} and \code{clusters}.
These rows are removed. Thus, \code{tx.group} and \code{clusters} should only be
supplied when intending to use them in \code{effect_robustness} or \code{dif_report}.
}
\examples{
measure <- data.frame(tx = rep(c("tx", "control"), times = 10),
                      gender = rep(c("male", "female"), each = 10),
                      item1 = sample(c(0,1), 20, replace = TRUE),
                      item2 = sample(c(0,1), 20, replace = TRUE),
                      item3 = sample(c(0,1), 20, replace = TRUE),
                      item4 = sample(c(0,1), 20, replace = TRUE),
                      item5 = sample(c(0,1), 20, replace = TRUE))
## add missing data
measure <- measure`[`c(1, 4), `]` <- NA
measure <- measure`[`c(2, 5), 1`]` <- NA

dif_prep(measure`[`, -c(1, 2)`]`,
         dif.group = measure$gender,
         tx.group = measure$tx,
         na0 = TRUE)

}
