% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Report.R
\name{dif_report}
\alias{dif_report}
\title{Generate DIF analysis report}
\usage{
dif_report(
  dif.analysis = NULL,
  dif.models = NULL,
  effect.robustness = NULL,
  report.type = "dif.effects",
  report.title = NULL,
  measure.name = NULL,
  file.name = measure.name,
  dataset.name = NULL,
  bias.method = "IRT"
)
}
\arguments{
\item{dif.analysis}{An object returned from \code{dif_analysis}}

\item{report.type}{Produce report including both the DIF analysis results and treatment effect robustness checks ("dif.effects"; default), only DIF analysis ("dif.only"), or only treatment effect robustness checks ("effects.only")?}

\item{report.title}{Title of the report}

\item{measure.name}{Name of the measure being evaluated for DIF, which is used in the report.}

\item{file.name}{File name to create on disk. Defaults to \code{measure.name}. The file path can also be specified here. The file is saved to the working directory by default.}

\item{dataset.name}{(optional) name of the dataset where the measure and
item responses came from, which is printed in the report. For World Bank data, this will typically be a country.}

\item{bias.method}{From which method in \code{dif.analysis} should the biased items, if detected, be extracted? The options are "MH", "logistic", or "IRT" (default).}

\item{irt.scoring}{Factor score estimation method, which is passed to
\code{\link[mirt]{fscores}}. Default is "WLE". See \code{\link[mirt]{fscores}} documentation for other options.}
}
\value{
Uses the template "Bias_Correction_Report.Rmd" to produce a
report summarizing whether any items on \code{measure.name} are biased with
respect to \code{dif.group.name}, and, if so, to what extent this affects
treatment contrasts on \code{measure.name}.
}
\description{
Produces a report summarizing an analysis of measurement bias for a given
measure and grouping variable
}
\details{
With \code{bias.method = "IRT"}, the direction of uniform DIF is determined using
the b parameter for dichotomous items and the location of the first threshold for
polytomous items.
}
\examples{
data("mdatlang")

# prep data
dif.data <- dif_data_prep(item.data = mdatlang`[`5:ncol(mdatlang)],
                             tx.group.id = mdatlang$treated,
                             dif.group.id = mdatlang$gender,
                             cluster.id = mdatlang$clusterid,
                             na.to.0 = TRUE)

# DIF analysis by dif.group.id
# using rest scores and binning match.scores by deciles to avoid empty cells in MH analysis
dif.analysis <- dif_analysis(dif.data = dif.data,
                           methods =  c("MH", "IRT")
                           match.type = "Rest",
                           match.bins = seq(0, 1, by = .1))
}
