% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Report.R
\name{dif_report}
\alias{dif_report}
\title{Produces a report (via .Rmd) that summarizes. measurement bias analysis for a given
scale and given grouping variable}
\usage{
dif_report(
  dif.results,
  dataset.name,
  measure.name,
  comparison.name = "Treatment Condition",
  bias.method = "IRT",
  irt.method = "WLE",
  conditional = NULL
)
}
\arguments{
\item{dif.results}{desc}

\item{dataset.name}{desc}

\item{measure.name}{desc}

\item{comparison.name}{desc}

\item{bias.method}{desc}

\item{irt.method}{desc}

\item{conditional}{desc}
}
\value{
Uses the template file "Bias_Correction_Report.Rmd" to produce a report summarizing whether any items on `measure.name` are biased with respect to `comparison.name`, and, if so, to what extent this affects treatment contrasts on `measure.name`.
}
\description{
Produces a report (via .Rmd) that summarizes. measurement bias analysis for a given
scale and given grouping variable
}
\note{
If `conditional = NULL` the grouping variable given in `comparison.name` is taken to be the treatment constrast. If `conditional = some.variable` ... actually this is kind of complicated.
}
