% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Report.R, R/newDIF-Report.R
\name{dif_report}
\alias{dif_report}
\title{Generate DIF analysis report}
\usage{
dif_report(
  dif.analysis = NULL,
  dif.models = NULL,
  effect.robustness = NULL,
  report.type = "dif.effects",
  report.title = NULL,
  measure.name = NULL,
  file.name = measure.name,
  dataset.name = NULL,
  bias.method = "IRT"
)

dif_report(
  dif.analysis = NULL,
  dif.models = NULL,
  effect.robustness = NULL,
  report.type = "dif.effects",
  report.title = NULL,
  measure.name = NULL,
  file.name = measure.name,
  dataset.name = NULL,
  bias.method = "IRT"
)
}
\arguments{
\item{dif.analysis}{An object returned from \code{dif_analysis}}

\item{report.title}{Title of the report}

\item{measure.name}{Name of the measure being evaluated for DIF, which is used in the report.}

\item{file.name}{File name to create on disk. Defaults to \code{measure.name}. The file path can also be specified here. The file is saved to the working directory by default.}

\item{dataset.name}{(optional) name of the dataset where the measure and
item responses came from, which is printed in the report. For World Bank data, this will typically be a country.}

\item{bias.method}{From which method in \code{methods} should the biased items, if detected, be extracted? The options are "MH", "logistic", or "IRT" (default).}

\item{dif.group.name}{name of the group used in the DIF analysis, which is
used in the report.}

\item{tx.group}{When `NULL` (default) unconditional treatment effects are assessed
for robustness with the assumption that the treatment indicator was used as the
grouping variable in \code{dif_analysis}. If treatment indicator was not the grouping
variable in \code{dif_analysis}, the factor vector for treatment indicator needs to be
supplied here. Robustness of the conditional treatment effects is then assessed.}

\item{methods}{A character \code{vector} with one or more of \code{c("loess", "MH", "logistic", "IRT")}. The default is all four methods.}

\item{irt.scoring}{Factor score estimation method, which is passed to
\code{\link[mirt]{fscores}}. Default is "WLE". See \code{\link[mirt]{fscores}} documentation for other options.}
}
\value{
Uses the template "Bias_Correction_Report.Rmd" to produce a
report summarizing whether any items on \code{measure.name} are biased with
respect to \code{dif.group.name}, and, if so, to what extent this affects
treatment contrasts on \code{measure.name}.

Uses the template "Bias_Correction_Report.Rmd" to produce a
report summarizing whether any items on \code{measure.name} are biased with
respect to \code{dif.group.name}, and, if so, to what extent this affects
treatment contrasts on \code{measure.name}.
}
\description{
Produces a report summarizing an analysis of measurement bias for a given
measure and grouping variable

Produces a report summarizing an analysis of measurement bias for a given
measure and grouping variable
}
\details{
With \code{bias.method = "IRT"}, the direction of uniform DIF is determined using
the b parameter for dichotomous items and the location of the first threshold for
polytomous items.

With \code{bias.method = "IRT"}, the direction of uniform DIF is determined using
the b parameter for dichotomous items and the location of the first threshold for
polytomous items.
}
\examples{
wb.measure <- data.frame(tx = rep(c("tx", "control"), times = 10),
                         gender = rep(c("male", "female"), each = 10),
                         item1 = sample(c(0,1), 20, replace = TRUE),
                         item2 = sample(c(0,1), 20, replace = TRUE),
                         item3 = sample(c(0,1), 20, replace = TRUE),
                         item4 = sample(c(0,1), 20, replace = TRUE),
                         item5 = sample(c(0,1), 20, replace = TRUE))

## Unconditional Treatment Effects
# DIF analysis by treatment condition
dif.by.tx <- dif_analysis(measure.data = wb.measure[3:7],
                          dif.group = wb.measure$tx,
                          score.type = "Rest",
                          methods = c("loess", "logistic"))

dif_report(dif.analysis = dif.by.tx,
           file.name = "WB-LR.html",
           measure.name = "Letter Recognition",
           dif.group.name = "Treatment Condition",
           bias.method = "logistic",
           irt.scoring = "WLE")

## Conditional Treatment Effects
# DIF analysis by gender
dif.by.gender <- dif_analysis(measure.data = wb.measure`[`3:7`]`,
                              dif.group = wb.measure$gender,
                              score.type = "Rest",
                              methods = c("loess", "logistic"))

dif_report(dif.analysis = dif.by.gender,
           file.name = "WB-LR.html",
           measure.name = "Letter Recognition",
           dif.group.name = "Gender",
           bias.method = "logistic",
           irt.scoring = "WLE",
           tx.group = wb.measure$tx)

data("mdatlang")

# prep data
dif.data <- dif_data_prep(item.data = mdatlang`[`5:ncol(mdatlang)],
                             tx.group.id = mdatlang$treated,
                             dif.group.id = mdatlang$gender,
                             cluster.id = mdatlang$clusterid,
                             na.to.0 = TRUE)

# DIF analysis by dif.group.id
# using rest scores and binning match.scores by deciles to avoid empty cells in MH analysis
dif.analysis <- dif_analysis(dif.data = dif.data,
                           methods =  c("MH", "IRT")
                           match.type = "Rest",
                           match.bins = seq(0, 1, by = .1))
}
