% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIF-Report.R
\name{dif_report}
\alias{dif_report}
\title{Generate DIF analysis report}
\usage{
dif_report(
  dif.analysis,
  file.name,
  measure.name,
  dif.group.name,
  dataset.name = NULL,
  report.title = NULL,
  bias.method = "IRT",
  irt.scoring = "WLE",
  tx.group = NULL
)
}
\arguments{
\item{dif.analysis}{an object returned from \code{dif_analysis}}

\item{file.name}{file name to create on disk.}

\item{measure.name}{name of the measure being evaluated for DIF, which is
used in the report.}

\item{dif.group.name}{name of the group used in the DIF analysis, which is
used in the report.}

\item{dataset.name}{(optional) name of the dataset where the measure and
item responses came from, which is printed in the report.
For World Bank data, this will typically be a country.}

\item{report.title}{title of the report}

\item{bias.method}{options are "MH", "logistic", or "IRT" (default).
Which method in the \code{dif_analysis} object should the biased items, if detected, be
extracted?}

\item{irt.scoring}{factor score estimation method, which is passed to
\code{\link[mirt]{fscores}}. Default is "WLE". See \code{\link[mirt]{fscores}}
documentation for other options.}

\item{tx.group}{When `NULL` (default) unconditional treatment effects are assessed
for robustness with the assumption that the treatment indicator was used as the
grouping variable in \code{dif_analysis}. If treatment indicator was not the grouping
variable in \code{dif_analysis}, the factor vector for treatment indicator needs to be
supplied here. Robustness of the conditional treatment effects is then assessed.}
}
\value{
Uses the template "Bias_Correction_Report.Rmd" to produce a
report summarizing whether any items on \code{measure.name} are biased with
respect to \code{dif.group.name}, and, if so, to what extent this affects
treatment contrasts on \code{measure.name}.
}
\description{
Produces a report summarizing an analysis of measurement bias for a given
scale and grouping variable
}
\details{
With \code{bias.method = "IRT"}, the direction of uniform DIF is determined using
the b parameter for dichotomous items and the location of the first threshold for
polytomous items.
}
\examples{
wb.measure <- data.frame(tx = rep(c("tx", "control"), times = 10),
                         gender = rep(c("male", "female"), each = 10),
                         item1 = sample(c(0,1), 20, replace = TRUE),
                         item2 = sample(c(0,1), 20, replace = TRUE),
                         item3 = sample(c(0,1), 20, replace = TRUE),
                         item4 = sample(c(0,1), 20, replace = TRUE),
                         item5 = sample(c(0,1), 20, replace = TRUE))

## Unconditional Treatment Effects
# DIF analysis by treatment condition
dif.by.tx <- dif_analysis(measure.data = wb.measure[3:7],
                          dif.group = wb.measure$tx,
                          score.type = "Rest",
                          methods = c("loess", "logistic"))

dif_report(dif.analysis = dif.by.tx,
           file.name = "WB-LR.html",
           measure.name = "Letter Recognition",
           dif.group.name = "Treatment Condition",
           bias.method = "logistic",
           irt.scoring = "WLE")

## Conditional Treatment Effects
# DIF analysis by gender
dif.by.gender <- dif_analysis(measure.data = wb.measure`[`3:7`]`,
                              dif.group = wb.measure$gender,
                              score.type = "Rest",
                              methods = c("loess", "logistic"))

dif_report(dif.analysis = dif.by.gender,
           file.name = "WB-LR.html",
           measure.name = "Letter Recognition",
           dif.group.name = "Gender",
           bias.method = "logistic",
           irt.scoring = "WLE",
           tx.group = wb.measure$tx)

}
