% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robustness-Functions.R
\name{effect_robustness}
\alias{effect_robustness}
\title{Robustness of treatment effects}
\usage{
effect_robustness(
  scale.data,
  dif.group,
  biased.items = NULL,
  no.var.items = integer(),
  no.var.by.group.items = integer(),
  poly = integer(),
  no.dif.mod = NULL,
  irt.scoring = "WLE",
  tx.group = NULL,
  clusters = NULL
)
}
\arguments{
\item{scale.data}{data frame of dichotomous item responses with subjects in rows
and items in columns}

\item{dif.group}{factor vector of group membership for which DIF is evaluated}

\item{biased.items}{integer vector; location of biased items in \code{scale.data}}

\item{no.var.items}{integer vector; location of items in \code{scale.data} with 0 variance}

\item{no.var.by.group.items}{integer vector; location of items in \code{scale.data}
with 0 variance within a \code{dif.group}}

\item{poly}{integer vector; location of polytomous items in \code{scale.data}}

\item{no.dif.mod}{\code{MultipleGroupClass} object from \code{mirt} where parameters are
constrained to be equal between levels of \code{dif.group}}

\item{irt.scoring}{factor score estimation method, which is passed to the
\code{method} argument of \code{\link[mirt]{fscores}}}

\item{tx.group}{factor vector for treatment indicator}

\item{clusters}{vector indicating cluster membership}

\item{score}{vector of total scores}
}
\value{
\code{effect_robustness} - a two-item list containing 1) data.frame of treatment effect
estimates and 2) character string of the groups being compared \cr
\code{smd_wrapper} - numeric value
}
\description{
Estimates treatment effects and robustness in the presence of DIF
}
\details{
Treatment effects are estimated as the standardized mean difference.
If \code{tx.group} is specified, the treatment effect is estimated for each group in
\code{dif.group} as well as the \code{tx.group} by \code{dif.group} interaction.
The interaction is estimated as difference in within-dif.group treatment effects
divided by the pooled SD of the dif.group1 and dif.group2 control groups.

The reported total score reliability is \eqn{\alpha}.
}
