% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Robustness-Functions.R
\name{get_robustness}
\alias{get_robustness}
\alias{calc_smd}
\alias{smd_wrapper}
\title{Robustness of treatment effects}
\usage{
get_robustness(
  scale.data,
  dif.group,
  biased.items = NULL,
  no.var.items = integer(),
  no.var.by.group.items = integer(),
  poly = integer(),
  no.dif.mod = NULL,
  irt.scoring = "WLE",
  tx.group = NULL
)

calc_smd(m1, m2, sd1, sd2 = NULL, n1, n2, sample = "ind")

smd_wrapper(score, dif.group, tx.group = NULL)
}
\arguments{
\item{scale.data}{data frame of dichotomous item responses with subjects in rows
and items in columns}

\item{dif.group}{factor vector of group membership for which DIF is evaluated}

\item{biased.items}{integer vector; location of biased items in \code{scale.data}}

\item{no.var.items}{integer vector; location of items in \code{scale.data} with 0 variance}

\item{no.var.by.group.items}{integer vector; location of items in \code{scale.data}
with 0 variance within a \code{dif.group}}

\item{poly}{integer vector; location of polytomous items in \code{scale.data}}

\item{no.dif.mod}{\code{MultipleGroupClass} object from \code{mirt} where parameters are
constrained to be equal between levels of \code{dif.group}}

\item{irt.scoring}{factor score estimation method, which is passed to the
\code{method} argument of \code{\link[mirt]{fscores}}}

\item{tx.group}{factor vector for treatment indicator}

\item{m1}{mean for dif.group 1}

\item{m2}{mean for dif.group 2}

\item{sd1}{sd for dif.group 1}

\item{sd2}{sd for dif.group 2}

\item{n1}{sample size for dif.group 1}

\item{n2}{sample size for dif.group 2}

\item{score}{vector of scale scores}
}
\value{
\code{get_robustness} - a two-item list containing 1) data.frame of treatment effect estimates
and 2) character string of the groups being compared \cr
\code{calc_smd} - numeric value
\cod{smd_wrapper} - numeric value
}
\description{
Calculates treatment effect and robustness in the presence of DIF
}
\details{

}
