% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper-Functions.R
\name{sum_score}
\alias{sum_score}
\alias{extract_bi}
\alias{format_flex}
\alias{est_alpha}
\alias{coeff_alpha}
\title{Internal convenience functions}
\usage{
sum_score(item.data, drops = NULL, poly.items = integer())

extract_bi(dif.analysis)

format_flex(df, bold.bias = "no", digits = 3)

est_alpha(item.data)

coeff_alpha(dif.models, std.group = NULL)
}
\arguments{
\item{item.data}{frame of item responses with subjects in rows
and items in columns.}

\item{drops}{integer vector; item locations in \code{item.data} to exclude from
score calculation}

\item{poly.items}{integer vector; location of polytomous items in \code{item.data}}

\item{dif.analysis}{an object returned from \code{\link[WBdif]{dif_analysis}}}

\item{df}{a data frame}

\item{bold.bias}{character indicating whether results tables should bold instances
of DIF or not ("no", the default). If so, must specify whether the table
displays model comparisons ("model") or item-level tests ("item").}

\item{digits}{integer; number decimal places to print in report tables}
}
\value{
\code{sum_score} returns a numeric vector of summed scores \cr
\code{extract_bi} returns list of biased items from each \code{dif.analysis} method \cr
\code{format_flex} returns a \code{\link[flextable]{flextable}} object \cr
\code{est_alpha} returns a numeric value
}
\description{
Collection of convenience functions to improve automaticity and code readability
}
\details{
\code{sum_score} sums item responses into a scale score. Polytomous items are
unit scaled so each item has a maximum value of 1. \cr
\code{extract_bi} extracts the biased items identified from each \code{dif.analysis}
method. If no biased items were detected, a character string indicated so is extracted. \cr
\code{format_flex} converts the DIF results from a data frame to a
\code{\link[flextable]{flextable}} and formats the table for printing in the report. \cr
}
